fastapi/dockerfiles/requirements.txt
fastapi~=0.109
uvicorn~=0.27
boto3~=1.34
mlflow~=2.10
numpy~=1.24
pandas~=2.0
scikit-learn~=1.3
protobuf==4.23.3
grpcio==1.57.0
grpcio-tools==1.57.0
strawberry-graphql~=0.255.0